Remove gap prop from Grid component
Is not possible to use gap without causing wrap component content, the ways to handle this means adding more safelists, so better use a div with mx-value on app host

The Grid component and its documentation have been updated to remove support for the 'gap' prop. All related code, examples, and Tailwind safelist entries for gap classes have been deleted to simplify the API and usage.